Rename raft_movement to raft_utils
